+++ packages/components/CHANGELOG.md
@@ -1,4 +1,37 @@
+### 📈 Stats
+* Uncompressed: 439 KB (+29 KB / +7%)
+* Gzipped: 132 KB (+7 KB / +6%)
+
+### 📈 Stats
+
+## 3.23.0
+
+### Minor Changes
+
+- Add sd-combobox and improve sd-select. _[`#1633`](https://github.com/solid-design-system/solid/pull/1633) [`8a16da8`](https://github.com/solid-design-system/solid/commit/8a16da830e7fdd135d9dcbbc540c48ef567dbe02) [@DanielHargesheimer](https://github.com/DanielHargesheimer)_
+
+  - add configurable `sd-combobox` component
+  - add keyboard navigation and accessibility features
+  - add multi-select and tag support
+  - use backspace key to remove latest selected option on `sd-combobox`
+  - use backspace key to remove focused tag and related option
+  - improve tags presentation
+  - show tags in selected order
+  - add sd-optgroup component for better grouping of options
+
+* Uncompressed: 410 KB (+1 KB / +0%)
+* Gzipped: 125 KB (unchanged)
+
 # @solid-design-system/components
+
+## 3.22.18
+
+### Patch Changes
+
+- Improved sd-select a11y: _[`#1710`](https://github.com/solid-design-system/solid/pull/1710) [`ab76208`](https://github.com/solid-design-system/solid/commit/ab76208754415a3c146795b7e8329df3006ea74c) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+  - Implemented `aria-invalid` attribute to semantically indicate component validity.
+  - Implemented live announcement to screen readers when a tag is removed.
 
 ## 3.22.17